refactor: rename utils file to 'utils'
